fix(dashboard): improve config schema UX — ranges, selects, i18n (#2644)
* fix(dashboard): improve config schema UX — ranges, selects, i18n labels

- Add min/max/step to all number fields across every config section
- Change fixed-value fields to select (language, embedding_provider,
  image_provider) with { value, label } i18n support in ConfigPage
- Introduce number_select type for interval fields (consolidation,
  cleanup) with translated preset labels via i18n keys
- Fix decay_rate f32→f64 to eliminate floating-point display artifacts
- Show workspaces_dir effective path instead of null when unset
- Sync registry provider files on update, not just on first copy
- Show generating placeholder in ChatInput while response is loading
- Add missing desc_*/fld_* i18n keys for all new constrained fields

* feat(dashboard): i18n labels for all select options via key lookup

* fix(providers): add LocalOffline status to prevent offline local providers from reappearing

Introduce AuthStatus::LocalOffline so the async health probe and
detect_auth() no longer fight over local provider status.

Previously, the probe would mark unreachable vllm/lmstudio as Missing,
but any subsequent detect_auth() call (triggered by saving API keys,
config changes, etc.) would reset Missing back to NotRequired, causing
the providers to reappear in the chat model switcher even when not
running.

Fix: probe sets LocalOffline on failure; detect_auth() skips providers
already in LocalOffline state. Also update sync_flat_files to overwrite
existing builtin provider TOMLs so new fields (e.g. supports_thinking)
stay in sync with the registry.

* fix(dashboard): refresh agent data after /model slash command

* feat(dashboard): slash command autocomplete with keyboard navigation and /model completion

- Replace old dropdown with new one showing activeIndex highlight and arg hints
- Add /model second-level completion: fetches model list from API, filters by typed arg
- Keyboard navigation: ArrowUp/Down to move selection, Enter/Tab to confirm, Escape to clear
- noArgs commands fill+send immediately; arg commands (model, verbose) fill input with trailing space
- Fix useQuery type error by wrapping listModels in arrow function

* fix(dashboard): resolve TypeScript unused variable and type errors in AgentsPage and ProvidersPage

* fix(dashboard): render multi-line system messages as markdown card, /help as formatted table

* feat(config): expand language select to 18 major languages

* fix(config): correct save status check — server returns applied/applied_partial not ok/saved

* fix(dashboard): system message multi-line layout keeps original divider aesthetic

+++ crates/librefang-types/src/config/types.rs
@@ -3802,7 +3802,7 @@ pub struct MemoryConfig {
     /// Maximum memories before consolidation is triggered.
     pub consolidation_threshold: u64,
     /// Memory decay rate (0.0 = no decay, 1.0 = aggressive decay).
-    pub decay_rate: f32,
+    pub decay_rate: f64,
     /// Embedding provider. Valid values: `"openai"`, `"groq"`, `"mistral"`,
     /// `"together"`, `"fireworks"`, `"cohere"`, `"ollama"`, `"bedrock"`,
     /// `"vllm"`, `"lmstudio"`, or `"auto"`.
